requirements.txt
# Core ML/AI libraries
faiss-cpu>=1.7.4
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
POT>=0.9.1
scikit-learn>=1.3.0
packaging>=21.0

# Data handling and processing
datasets>=2.14.0
pandas>=1.5.0
pyyaml>=5.1
tqdm>=4.66.0

# Advanced NLP and mathematical foundations
spacy[en_core_web_sm]>=3.7.0
nltk>=3.8.0
# gensim>=4.3.0  # Temporarily disabled due to build issues with Python 3.13
pingouin>=0.5.0
statsmodels>=0.14.0
networkx>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Evaluation and benchmarking
beir>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
jupyter>=1.0.0
pre-commit>=3.6.0

# Optimization and constraint solving
z3-solver>=4.12.0

# Data serialization and validation
msgspec>=0.18.0
pydantic>=2.0.0
orjson>=3.8.0
packaging>=21.0

# Security and hashing
blake3>=0.3.3

# Cloud services (Google Cloud)
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.11.0
google-cloud-vision>=3.4.0

# Workflow orchestration  
# airflow>=2.7.0  # Temporarily disabled - requires Python <3.12

# System monitoring and utilities
psutil>=5.9.0
GitPython>=3.1.0

# Service discovery and configuration
# python-etcd3>=0.12.0  # Temporarily disabled - not available on Python 3.13
python-consul>=1.1.0

# Async and HTTP libraries
aioredis>=2.0.0
httpx>=0.24.0
redis>=4.5.0

# NLP processing and text handling
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
PyPDF2>=3.0.1
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0
camelot-py>=0.11.0
ftfy>=6.1.0
unidecode>=1.3.0
pyspellchecker>=0.7.0
langdetect>=1.0.9
textstat>=0.7.0

# Process mining and genetic algorithms
pm4py>=2.7.0
deap>=1.4.0

# Information retrieval
# pyserini>=0.22.0  # Commented out due to Java dependency
whoosh>=2.7.4

# Machine learning extensions
ray>=2.7.0
torch-geometric>=2.3.0
# node2vec>=0.4.0  # Commented out due to gensim build issues on Python 3.13

# Search and indexing
elasticsearch>=8.9.0

# Parallel processing and workflow
joblib>=1.3.0
dask>=2023.8.0

# Database connectivity
sqlalchemy>=2.0.0

# Text processing
lark>=1.1.0

# Validation and serialization
jsonschema>=4.19.0
toml>=0.10.2

# Monitoring and observability
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
loguru>=0.7.0
rich>=13.5.0

# Data analysis
tabula-py>=2.8.0

# Web scraping and data extraction  
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Container orchestration
kubernetes>=27.2.0

# Message queuing
kafka-python>=2.0.2

# Data validation
hypothesis>=6.82.0

# Additional utilities
setuptools>=68.0.0
pillow>=10.0.0

# Topological Data Analysis libraries (optional)
# gudhi>=3.7.0  # Install with: conda install -c conda-forge gudhi
# ripser>=0.6.0  # Install with: pip install ripser